+++ 404.php
@@ -1,4 +1,4 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit;
 get_header(); ?>
 <?php do_action('totalpress_404_start') ;?>