+++ assets/customizer/inc/totalpress-header-sidebar-section.php
@@ -1,7 +1,7 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit;
 TotalPress_Kirki::add_section( 'totalpress_header_sidebar_layout_section', array(
-    'title'       => esc_html__('Header Sidebar Layout','totalpress'),
+    'title'       => esc_html__('Layout','totalpress'),
     'panel'       => 'totalpress_header_sidebar_options',
     'priority'    => 1,
     'capability'  => 'edit_theme_options',
@@ -10,7 +10,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'radio-buttonset',
 	'settings'    => 'totalpress_head_sidebar_alignment',
 	'section'     => 'totalpress_header_sidebar_layout_section',
-	'label'       => esc_html__('Header Sidebar Alignment','totalpress'),
+	'label'       => esc_html__('Alignment','totalpress'),
 	'tooltip'     => esc_html__('Set the alignment of the Header sidebar. Default is Right.','totalpress'),
 	'default'     => 'right',
 	'priority'    => 4,