+++ assets/customizer/inc/totalpress-layout-section.php
@@ -1,5 +1,23 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit;
+TotalPress_Kirki::add_field( 'theme_customizer', array(
+	'type'        => 'select',
+	'settings'    => 'totalpress_blog_layout',
+	'section'     => 'totalpress_layout_options',
+	'label'       => esc_html__('Theme Layout','totalpress'),
+	'tooltip'     => esc_html__('Set initial layout for the theme. Will also affect the layout of posts, pages, archives, categories etc.','totalpress'),
+	'default'     => 'right_sidebar',
+	'priority'    => 1,
+	'transport'   => 'auto',
+	'choices'     => array(
+		'right_sidebar'   => esc_html__('Content | Sidebar','totalpress'),
+		'left_sidebar'    => esc_html__('Sidebar | Content','totalpress'),
+		'sidebars_right'  => esc_html__('Content | Sidebar | Sidebar','totalpress'),
+		'sidebars_left'   => esc_html__('Sidebar | Sidebar | Content','totalpress'),
+		'both_sidebars'   => esc_html__('Sidebar | Content | Sidebar','totalpress'),
+		'no_sidebars'     => esc_html__('No Sidebars (Full Width)','totalpress'),
+	),
+));
 TotalPress_Kirki::add_section( 'totalpress_layout_options', array(
     'title'       => esc_html__('Layout','totalpress'),
     'panel'       => 'totalpress_general_options',
@@ -13,7 +31,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'tooltip'     => esc_html__('Set the width of Foundations&#39; main grid container. Default is 1200px (75rem).','totalpress'),
 	'section'     => 'totalpress_layout_options',
 	'default'     => '1200px',
-	'priority'    => 1,
+	'priority'    => 2,
 	'transport'   => 'auto',
 	'output'      => array(
 		array(
@@ -21,24 +39,6 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 			'function' => 'css',
 			'property' => 'max-width',
 		),
-	),
-));
-TotalPress_Kirki::add_field( 'theme_customizer', array(
-	'type'        => 'select',
-	'settings'    => 'totalpress_blog_layout',
-	'section'     => 'totalpress_layout_options',
-	'label'       => esc_html__('Theme Layout','totalpress'),
-	'tooltip'     => esc_html__('Set initial layout for the theme. Will also affect the layout of posts, pages, archives, categories etc.','totalpress'),
-	'default'     => 'right_sidebar',
-	'priority'    => 2,
-	'transport'   => 'auto',
-	'choices'     => array(
-		'right_sidebar'   => esc_html__('Content | Sidebar','totalpress'),
-		'left_sidebar'    => esc_html__('Sidebar | Content','totalpress'),
-		'sidebars_right'  => esc_html__('Content | Sidebar | Sidebar','totalpress'),
-		'sidebars_left'   => esc_html__('Sidebar | Sidebar | Content','totalpress'),
-		'both_sidebars'   => esc_html__('Sidebar | Content | Sidebar','totalpress'),
-		'no_sidebars'     => esc_html__('No Sidebars (Full Width)','totalpress'),
 	),
 ));
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(