+++ assets/customizer/inc/totalpress-navigation-section.php
@@ -1,7 +1,7 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit;
 TotalPress_Kirki::add_section( 'totalpress_navigation_section', array(
-    'title'       => esc_html__('Navigation Layout','totalpress'),
+    'title'       => esc_html__('Layout','totalpress'),
     'panel'       => 'totalpress_navigation_options',
     'priority'    => 4,
     'capability'  => 'edit_theme_options',
@@ -9,7 +9,7 @@ TotalPress_Kirki::add_section( 'totalpress_navigation_section', array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'select',
 	'settings'    => 'totalpress_nav_container',
-	'label'       => __('Main Navigation Width','totalpress' ),
+	'label'       => __('Main Container Width','totalpress' ),
 	'tooltip'     => esc_html__('Select &quot;Contain to Grid&quot; to keep the main menu container within the grid container. Select &quot;Full Width&quot; have the main menu container expand 100% to fill the screen. Default is &quot;Contain to Grid&quot;','totalpress'),
 	'section'     => 'totalpress_navigation_section',
 	'default'     => 'contain',
@@ -22,7 +22,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'select',
 	'settings'    => 'totalpress_inner_nav_container',
-	'label'       => __('Inner Main Menu Width','totalpress' ),
+	'label'       => __('Inner Container Width','totalpress' ),
 	'tooltip'     => esc_html__('Select &quot;Contain to Grid&quot; to keep the main menu contained within the grid container. Select &quot;Full Width&quot; to have the main menu expand 100% to fill the screen. Default is &quot;Contain to Grid&quot;','totalpress'),
 	'section'     => 'totalpress_navigation_section',
 	'default'     => 'contain',
@@ -35,7 +35,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 TotalPress_Kirki::add_field( 'theme_customizer', array(
 	'type'        => 'select',
 	'settings'    => 'totalpress_nav_position',
-	'label'       => esc_html__('Main Menu Position','totalpress'),
+	'label'       => esc_html__('Menu Position','totalpress'),
 	'tooltip'     => esc_html__('Select where you want to display the Main Menu.','totalpress'),
 	'section'     => 'totalpress_navigation_section',
 	'default'     => 'bottom_of_header',
@@ -49,7 +49,7 @@ TotalPress_Kirki::add_field( 'theme_customizer', array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array( //control
 	'type'        => 'radio-buttonset',
 	'settings'    => 'totalpress_theme_nav_alignment',
-	'label'       => esc_html__('Main Menu Alignment','totalpress'),
+	'label'       => esc_html__('Menu Alignment','totalpress'),
 	'tooltip'     => esc_html__('Set the alignment of the Main Menu.','totalpress'),
 	'section'     => 'totalpress_navigation_section',
 	'default'     => 'top-bar-left',