+++ assets/customizer/inc/totalpress-top-sidebar-section.php
@@ -1,7 +1,7 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit;
 TotalPress_Kirki::add_section( 'totalpress_top_sidebar_layout_section', array(
-    'title'       => esc_html__('Top Sidebar Layout','totalpress'),
+    'title'       => esc_html__('Layout','totalpress'),
     'panel'       => 'totalpress_top_sidebar_options',
     'priority'    => 1,
     'capability'  => 'edit_theme_options',
@@ -9,7 +9,7 @@ TotalPress_Kirki::add_section( 'totalpress_top_sidebar_layout_section', array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'select',
 	'settings'    => 'totalpress_top_sidebar_container',
-	'label'       => __('Main Top Sidebar Width','totalpress' ),
+	'label'       => __('Main Container Width','totalpress' ),
 	'tooltip'     => esc_html__('Select &quot;Contain to Grid&quot; to keep the main Top Sidebar contained within the grid container. Select &quot;Full Width&quot; have the main header expand 100% to fill the screen. Default is &quot;Contain to Grid&quot;','totalpress'),
 	'section'     => 'totalpress_top_sidebar_layout_section',
 	'default'     => 'contain',
@@ -23,7 +23,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'select',
 	'settings'    => 'totalpress_inner_top_sidebar_container',
-	'label'       => __('Inside Top Sidebar Width','totalpress' ),
+	'label'       => __('Inner Container Width','totalpress' ),
 	'tooltip'     => esc_html__('Select &quot;Contain to Grid&quot; to keep the inner Top Sidebar contained within the grid container. Select &quot;Full Width&quot; to have the inside header expand 100% to fill the screen. Default is &quot;Contain to Grid&quot;','totalpress'),
 	'section'     => 'totalpress_top_sidebar_layout_section',
 	'default'     => 'contain',
@@ -37,7 +37,7 @@ TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 TotalPress_Kirki::add_field('totalpress_theme_customizer',array(
 	'type'        => 'radio-buttonset',
 	'settings'    => 'totalpress_top_sidebar_alignment',
-	'label'       => esc_html__('Top SideBar Alignment','totalpress'),
+	'label'       => esc_html__('Alignment','totalpress'),
 	'tooltip'     => esc_html__('Set the alignment of the Top Sidebar area. Default is Right.','totalpress'),
 	'section'     => 'totalpress_top_sidebar_layout_section',
 	'default'     => 'right',