+++ page-templates/content-sidebar-sidebar.php
@@ -1,4 +1,4 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 /**
  * Template Name: Content | Sidebar | Sidebar
  * Template Post Type: post, page, aside, audio, chat, gallery, image, link, quote, status, video