+++ searchform.php
@@ -1,3 +1,3 @@
-<?php /* @version 1.0.19 */
+<?php /* @version 1.0.20 */
 if ( ! defined('ABSPATH')) exit; ?>
 <?php do_action('totalpress_search_form'); ?>